[25.09.24]: created a simple xcb window handling

+++ src/csc/pngine/window_handler/win32_handler/win32_handler.impl.cxx
@@ -12,7 +12,7 @@ class win32_handler_impl {
   HISTANCE pm_instance;
 
  public:
-  win32_handler_impl(uint32_t width, uint32_t height) {
+  win32_handler_impl(uint16_t width, uint16_t height) {
   }
 };
